# docs/backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "EvidenceBridge local demo API",
    "version": "1.0",
    "description": "One synthetic case. Human evidence review, not a hiring decision. Model keys remain server-side."
  },
  "components": {
    "securitySchemes": {
      "demoAdmin": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Demo-Admin-Token"
      }
    },
    "schemas": {
...
      "def-1": {
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "additionalProperties": false,
            "type": "object",
            "required": [
              "code",
              "message",
              "requestId",
              "retryable"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "requestId": {
                "type": "string"
              },
              "retryable": {
                "type": "boolean"
              }
            }
          }
        },
        "title": "ApiError"
      }
    }
  },
  "paths": {
...
    "/api/demo": {
      "get": {
        "summary": "Read the single shared case and current report",
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "4XX": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-1"
                }
              }
            }
          },
          "5XX": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-1"
                }
              }
            }
          }
        }
      }
    },
    "/api/demo/task/send": {
      "post": {
        "summary": "Send the preset task once",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "type": "object",
                "required": [
                  "schemaVersion",
                  "sessionId",
                  "taskId",
                  "datasetVersion",
                  "instructions"
                ],
                "properties": {
                  "schemaVersion": {
                    "type": "string",
                    "enum": [
                      "1.0"
                    ]
                  },
                  "sessionId": {
                    "minLength": 1,
                    "maxLength": 80,
                    "pattern": "^[A-Za-z0-9_-]+$",
                    "type": "string"
                  },
                  "taskId": {
                    "minLength": 1,
                    "maxLength": 80,
                    "pattern": "^[A-Za-z0-9_-]+$",
                    "type": "string"
                  },
                  "datasetVersion": {
                    "minLength": 1,
                    "maxLength": 80,
                    "pattern": "^[A-Za-z0-9_-]+$",
                    "type": "string"
                  },
                  "instructions": {
                    "minLength": 1,
                    "maxLength": 4000,
                    "pattern": "\\S",
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "minLength": 8,
              "maxLength": 100,
              "pattern": "^[A-Za-z0-9_-]+$",
              "type": "string"
            },
            "in": "header",
            "name": "idempotency-key",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "4XX": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-1"
                }
              }
            }
          },
          "5XX": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-1"
                }
              }
            }
          }
        }
      }
    },
    "/api/demo/submission": {
      "post": {
        "summary": "Save an immutable public work sample; private notes are excluded",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "type": "object",
                "required": [
                  "schemaVersion",
                  "sessionId",
                  "taskId",
                  "datasetVersion",
                  "candidateId",
                  "summary",
                  "findings",
                  "processEvidence"
                ],
                "properties": {
                  "schemaVersion": {
                    "type": "string",
                    "enum": [
                      "1.0"
                    ]
                  },
                  "sessionId": {
                    "minLength": 1,
                    "maxLength": 80,
                    "pattern": "^[A-Za-z0-9_-]+$",
                    "type": "string"
                  },
                  "taskId": {
                    "minLength": 1,
                    "maxLength": 80,
                    "pattern": "^[A-Za-z0-9_-]+$",
                    "type": "string"
                  },
                  "datasetVersion": {
                    "minLength": 1,
                    "maxLength": 80,
                    "pattern": "^[A-Za-z0-9_-]+$",
                    "type": "string"
                  },
                  "candidateId": {
                    "minLength": 1,
                    "maxLength": 80,
                    "pattern": "^[A-Za-z0-9_-]+$",
                    "type": "string"
                  },
                  "summary": {
                    "minLength": 1,
                    "maxLength": 8000,
                    "pattern": "\\S",
                    "type": "string"
                  },
                  "findings": {
                    "maxItems": 40,
                    "type": "array",
                    "items": {
                      "additionalProperties": false,
                      "type": "object",
                      "required": [
                        "id",
                        "section",
                        "title",
                        "detail",
                        "source",
                        "confidence"
                      ],
                      "properties": {
                        "id": {
                          "minLength": 1,
                          "maxLength": 80,
                          "pattern": "^[A-Za-z0-9_-]+$",
                          "type": "string"
                        },
                        "section": {
                          "anyOf": [
                            {
                              "type": "string",
                              "enum": [
                                "Key Findings"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "Hypotheses"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "Additional Evidence Needed"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "Recommended Next Steps"
                              ]
                            }
                          ]
                        },
                        "title": {
                          "minLength": 1,
                          "maxLength": 300,
                          "pattern": "\\S",
                          "type": "string"
                        },
                        "detail": {
                          "maxLength": 4000,
                          "type": "string"
                        },
                        "source": {
                          "maxLength": 80,
                          "type": "string"
                        },
                        "confidence": {
                          "anyOf": [
                            {
                              "type": "string",
                              "enum": [
                                "High"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "Medium"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "Low"
                              ]
                            }
                          ]
                        }
                      }
                    }
                  },
                  "processEvidence": {
                    "maxItems": 100,
                    "type": "array",
                    "items": {
                      "additionalProperties": false,
                      "type": "object",
                      "required": [
                        "id",
                        "at",
                        "title"
                      ],
                      "properties": {
                        "id": {
                          "minLength": 1,
                          "maxLength": 80,
                          "pattern": "^[A-Za-z0-9_-]+$",
                          "type": "string"
                        },
                        "at": {
                          "format": "date-time",
                          "type": "string"
                        },
                        "title": {
                          "minLength": 1,
                          "maxLength": 200,
                          "pattern": "\\S",
                          "type": "string"
                        },
                        "detail": {
                          "maxLength": 1000,
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "minLength": 8,
              "maxLength": 100,
              "pattern": "^[A-Za-z0-9_-]+$",
              "type": "string"
            },
            "in": "header",
            "name": "idempotency-key",
            "required": true
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "4XX": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-1"
                }
              }
            }
          },
          "5XX": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-1"
                }
              }
            }
          }
        }
      }
    },
    "/api/demo/analysis": {
      "post": {
        "summary": "Extract and validate observations from the exact saved submission",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "type": "object",
                "required": [
                  "schemaVersion",
                  "sessionId",
                  "taskId",
                  "datasetVersion",
                  "submissionId",
                  "contentFingerprint"
                ],
                "properties": {
                  "schemaVersion": {
                    "type": "string",
                    "enum": [
                      "1.0"
                    ]
                  },
                  "sessionId": {
                    "minLength": 1,
                    "maxLength": 80,
                    "pattern": "^[A-Za-z0-9_-]+$",
                    "type": "string"
                  },
                  "taskId": {
                    "minLength": 1,
                    "maxLength": 80,
                    "pattern": "^[A-Za-z0-9_-]+$",
                    "type": "string"
                  },
                  "datasetVersion": {
                    "minLength": 1,
                    "maxLength": 80,
                    "pattern": "^[A-Za-z0-9_-]+$",
                    "type": "string"
                  },
                  "submissionId": {
                    "minLength": 1,
                    "maxLength": 80,
                    "pattern": "^[A-Za-z0-9_-]+$",
                    "type": "string"
                  },
                  "contentFingerprint": {
                    "pattern": "^[a-f0-9]{64}$",
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "minLength": 8,
              "maxLength": 100,
              "pattern": "^[A-Za-z0-9_-]+$",
              "type": "string"
            },
            "in": "header",
            "name": "idempotency-key",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "202": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "4XX": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-1"
                }
              }
            }
          },
          "5XX": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-1"
                }
              }
            }
          }
        }
      }
    },
    "/api/demo/review": {
      "post": {
        "summary": "Save one human evidence decision; only the target requirement changes",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "type": "object",
                "required": [
                  "schemaVersion",
                  "sessionId",
                  "taskId",
                  "datasetVersion",
                  "submissionId",
                  "contentFingerprint",
                  "requirementId",
                  "decision",
                  "comment"
                ],
                "properties": {
                  "schemaVersion": {
                    "type": "string",
                    "enum": [
                      "1.0"
                    ]
                  },
                  "sessionId": {
                    "minLength": 1,
                    "maxLength": 80,
                    "pattern": "^[A-Za-z0-9_-]+$",
                    "type": "string"
                  },
                  "taskId": {
                    "minLength": 1,
                    "maxLength": 80,
                    "pattern": "^[A-Za-z0-9_-]+$",
                    "type": "string"
                  },
                  "datasetVersion": {
                    "minLength": 1,
                    "maxLength": 80,
                    "pattern": "^[A-Za-z0-9_-]+$",
                    "type": "string"
                  },
                  "submissionId": {
                    "minLength": 1,
                    "maxLength": 80,
                    "pattern": "^[A-Za-z0-9_-]+$",
                    "type": "string"
                  },
                  "contentFingerprint": {
                    "pattern": "^[a-f0-9]{64}$",
                    "type": "string"
                  },
                  "requirementId": {
                    "minLength": 1,
                    "maxLength": 80,
                    "pattern": "^[A-Za-z0-9_-]+$",
                    "type": "string"
                  },
                  "decision": {
                    "anyOf": [
                      {
                        "type": "string",
                        "enum": [
                          "confirm"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "needs_more_evidence"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "evidence_still_insufficient"
                        ]
                      }
                    ]
                  },
                  "comment": {
                    "minLength": 1,
                    "maxLength": 2000,
                    "pattern": "\\S",
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "minLength": 8,
              "maxLength": 100,
              "pattern": "^[A-Za-z0-9_-]+$",
              "type": "string"
            },
            "in": "header",
            "name": "idempotency-key",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "4XX": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-1"
                }
              }
            }
          },
          "5XX": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-1"
                }
              }
            }
          }
        }
      }
    },
    "/api/demo/reset": {
      "post": {
        "summary": "Reset synthetic case; old submission references become stale",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "type": "object",
                "required": [
                  "schemaVersion",
                  "sessionId"
                ],
                "properties": {
                  "schemaVersion": {
                    "type": "string",
                    "enum": [
                      "1.0"
                    ]
                  },
                  "sessionId": {
                    "minLength": 1,
                    "maxLength": 80,
                    "pattern": "^[A-Za-z0-9_-]+$",
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "minLength": 8,
              "maxLength": 100,
              "pattern": "^[A-Za-z0-9_-]+$",
              "type": "string"
            },
            "in": "header",
            "name": "idempotency-key",
            "required": true
          }
        ],
        "security": [
          {
            "demoAdmin": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "4XX": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-1"
                }
              }
            }
          },
          "5XX": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-1"
                }
              }
            }
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8787"
    }
  ]
}